+++ docs/11_0_0/gettingstarted/whatsnew.md
@@ -10,7 +10,7 @@ This page contains a list of big features. Please check the GitHub issues for al
   * Hotkey: added `bindMac` to use alternative binding on macOS.
   * Inplace: added `tabindex` for keyboard focus and toggle support.
   * InputText, InputTextarea: counter can count bytes instead of characters.
-  * OutputPanel: added `skeletons` facet to show while deferred loading is in progress.
+  * OutputPanel: added `skeleton` facet to show while deferred loading is in progress.
   * Spinner
     * Added buttons modes: `horizontal`, `horizontal-after` and `vertical`.
     * Added support to align input text using classes: `text-left`, `text-center`, `text-right`.